Improving test coverage
