+++ app/__init__.py
@@ -34,4 +34,7 @@ def create_app(config_name):
     from resources.users import api as reg_login
     api.add_namespace(reg_login, path='/api/v1')
 
+    from resources.drivers import api as driver
+    api.add_namespace(driver, path='/api/v1')
+
     return app
